+++ packages/vue-component-library/src/lib-components/BlockMediaWithText.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
 import { computed, defineAsyncComponent, ref } from 'vue'
 
-// Helper functions
-
 import type { PropType } from 'vue'
 import SvgIconHeadphones from 'ucla-library-design-tokens/assets/svgs/molecule-headphones.svg'
 import type { MediaGalleryItemType, MediaItemType } from '@/types/types'
+
+import { useTheme } from '@/composables/useTheme'
 
 const props = defineProps({
   sectionHeader: {
@@ -41,6 +41,13 @@ const props = defineProps({
     default: () => [],
   },
 })
+
+const theme = useTheme()
+
+const classes = computed(() => {
+  return ['media-with-text', theme?.value || '']
+})
+
 /* const SvgIconHeadphones = defineAsyncComponent(() =>
   import(
     'ucla-library-design-tokens/assets/svgs/molecule-headphones.svg'
@@ -49,8 +56,8 @@ const SvgIconPlayFilled = defineAsyncComponent(() =>
   import(
     'ucla-library-design-tokens/assets/svgs/icon-play-filled.svg'
   ))
-// const SmartLink = defineAsyncComponent(() =>
-// import('@/lib-components/SmartLink.vue'))
+const SmartLink = defineAsyncComponent(() =>
+  import('@/lib-components/SmartLink.vue'))
 const ButtonLink = defineAsyncComponent(() =>
   import('@/lib-components/ButtonLink.vue'))
 const MediaItem = defineAsyncComponent(() =>
@@ -91,18 +98,13 @@ const parsedIsDownload = computed(() => {
 </script>
 
 <template>
-  <div class="media-with-text">
+  <div :class="classes">
     <div class="text-grouping">
       <h3
         class="section-header"
         v-html="sectionHeader"
       />
       <div class="meta-mobile">
-        <!-- <smart-link
-          v-if="mediaLink"
-          :to="mediaLink"
-          class="media-link"
-        > -->
         <MediaItem
           v-if="item || coverImage"
           :item="item"
@@ -111,7 +113,6 @@ const parsedIsDownload = computed(() => {
           class="media-mobile"
           @click="showLightbox = true"
         />
-        <!-- </smart-link> -->
         <div
           v-if="!(item || coverImage)"
           class="no-media-mobile"
@@ -140,20 +141,30 @@ const parsedIsDownload = computed(() => {
         class="short-description"
         v-html="shortDescription"
       />
+
       <ButtonLink
-        v-if="buttonUrl"
+        v-if="buttonUrl && theme !== 'ftva'"
         class="button"
         :to="buttonUrl"
         :label="buttonText"
         :is-secondary="true"
         :is-download="parsedIsDownload"
       />
+      <SmartLink
+        v-else
+        :to="buttonUrl"
+        class="text-link"
+      >
+        {{ buttonText }}
+      </SmartLink>
     </div>
+
     <MediaItem
       v-if="item || coverImage"
       :item="item"
       :cover-image="coverImage"
       :cover-only="true"
+      :object-fit="theme === 'ftva' ? 'cover' : 'contain'"
       class="meta media"
       @click="isAudio || isVideo ? showLightbox = true : showLightbox = false"
     >
@@ -180,11 +191,13 @@ const parsedIsDownload = computed(() => {
         class="icon-headphones"
       />
     </MediaItem>
+
     <div
       v-if="!(item || coverImage)"
       class="no-media"
     />
     <!--eslint-disable-->
+
     <NewLightbox
       v-if="showLightbox"
       :items="lightboxItems"
@@ -195,247 +208,6 @@ const parsedIsDownload = computed(() => {
 </template>
 
 <style lang="scss" scoped>
-.media-with-text {
-  display: flex;
-  flex-direction: row;
-  flex-wrap: nowrap;
-  align-items: center;
-  justify-content: space-between;
-
-  max-width: $container-l-main + px;
-  // Themes
-  --background-color: var(--color-theme, var(--color-white));
-
-  &.color-grey {
-    --background-color: var(--color-secondary-grey-02);
-  }
-
-  .text-grouping {
-    margin-right: 50px;
-
-    display: flex;
-    flex-direction: column;
-    flex-wrap: nowrap;
-    align-content: flex-start;
-    align-items: flex-start;
-  }
-
-  .section-header {
-    @include step-3;
-    color: var(--color-primary-blue-03);
-    margin-bottom: 24px;
-  }
-
-  .short-description {
-    @include step-0;
-    margin-bottom: 24px;
-  }
-
-  .meta {
-    max-width: 500px;
-    z-index: 0;
-    position: relative;
-  }
-
-  .media {
-    z-index: 0;
-    position: relative;
-    max-width: 100%;
-    min-width: 426px;
-    height: auto;
-  }
-
-  .no-media {
-    z-index: 0;
-    position: relative;
-    width: 426px;
-    height: 240px;
-    background-color: var(--color-primary-blue-02);
-  }
-
-  .meta-mobile {
-    display: none;
-  }
-
-  .clippy {
-    z-index: 100;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-  }
-
-  .floating-highlight {
-    z-index: 200;
-    position: absolute;
-    bottom: 8px;
-    left: 5px;
-    width: 112px;
-    height: 72px;
-    background-color: var(--color-visit-fushia-03);
-
-    clip-path: polygon(0 0,
-        calc(100% - 37px) 0,
-        100% 75px,
-        calc(100% - 1.5px) 75px,
-        calc(100% - 38px) 1.5px,
-        0 1.5px);
-  }
-
-  .clipped-play {
-    bottom: 0;
-    left: 0;
-    z-index: 200;
-    position: absolute;
-    width: 112px;
-    height: 72px;
-    background-color: var(--background-color);
-
-    clip-path: polygon(0 0,
-        calc(100% - 38px) 0,
-        100% 78px,
-        calc(100% - 1.5px) 84px,
-        0 84px,
-        0 1.5px);
-  }
-
-  .svg__icon-play {
-    fill: var(--color-primary-blue-03);
-  }
-
-  .icon-play-filled {
-    z-index: 400;
-    position: absolute;
-    bottom: 16px;
-    margin-left: 24px;
-  }
-
-  .icon-headphones {
-    z-index: 400;
-    position: absolute;
-    left: 50%;
-    margin-left: -40px;
-    top: 50%;
-    margin-top: -40px;
-  }
-
-  // Breakpoints
-
-  @media #{$medium} {
-    .media {
-      width: calc(50% - 48px);
-      height: auto;
-      max-width: 100%;
-      min-width: 296px;
-    }
-
-    .no-media {
-      width: calc(50% - 48px);
-      max-width: 100%;
-      min-width: 296px;
-    }
-  }
-
-  @media #{$small} {
-    &.media-with-text {
-      display: flex;
-      flex-direction: column;
-      flex-wrap: nowrap;
-      justify-content: center;
-      align-content: center;
-      align-items: center;
-
-      max-width: 100%;
-
-      .text-grouping {
-        max-width: 100%;
-        margin-right: 0;
-      }
-
-      .meta {
-        display: none;
-      }
-
-      .meta-mobile {
-        display: block;
-
-        display: block;
-        width: 100%;
-        height: auto;
-        margin-bottom: 24px;
-        z-index: 0;
-        position: relative;
-      }
-
-      .media-mobile {
-        width: 100%;
-        height: auto;
-        z-index: 10;
-        position: relative;
-      }
-
-      .no-media-mobile {
-        width: 100%;
-        height: 200px;
-        z-index: 10;
-        position: relative;
-        background-color: var(--color-primary-blue-02);
-      }
-
-      .floating-highlight-mobile {
-        z-index: 200;
-        position: absolute;
-        bottom: 8px;
-        left: 5px;
-        width: 112px;
-        height: 72px;
-        background-color: var(--color-visit-fushia-03);
-
-        clip-path: polygon(0 0,
-            calc(100% - 37px) 0,
-            100% 75px,
-            calc(100% - 1.5px) 75px,
-            calc(100% - 38px) 1.5px,
-            0 1.5px);
-      }
-
-      .clipped-play-mobile {
-        bottom: 0;
-        left: 0;
-        z-index: 200;
-        position: absolute;
-        width: 112px;
-        height: 72px;
-        background-color: var(--background-color);
-
-        clip-path: polygon(0 0,
-            calc(100% - 38px) 0,
-            100% 78px,
-            calc(100% - 1.5px) 84px,
-            0 84px,
-            0 1.5px);
-      }
-
-      .icon-play-filled-mobile {
-        z-index: 400;
-        position: absolute;
-        bottom: 16px;
-        margin-left: 24px;
-      }
-
-      .icon-headphones-mobile {
-        z-index: 400;
-        position: absolute;
-        left: 50%;
-        margin-left: -40px;
-        top: 50%;
-        margin-top: -40px;
-      }
-
-      .button {
-        width: 100%;
-        margin-left: 0;
-      }
-    }
-  }
-}
+@import "@/styles/default/_block-media-with-text.scss";
+@import "@/styles/ftva/_block-media-with-text.scss";
 </style>