+++ packages/vue-component-library/src/lib-components/Flexible/MediaWithText.vue
@@ -4,6 +4,7 @@
 >
 import type { PropType } from 'vue'
 import { computed } from 'vue'
+import { useTheme } from '@/composables/useTheme'
 
 // COMPONENTS
 import BlockMediaWithText from '@/lib-components/BlockMediaWithText.vue'
@@ -37,10 +38,16 @@ const parsedContent = computed(() => {
     }
   })
 })
+
+const theme = useTheme()
+
+const classes = computed(() => {
+  return ['section-media-with-text', theme?.value || '']
+})
 </script>
 
 <template>
-  <div class="section-media-with-text">
+  <div :class="classes">
     <div class="flexible-page-block-container">
       <h2 class="visually-hidden">
         Resources
@@ -75,7 +82,7 @@ const parsedContent = computed(() => {
           class="flexible-media-with-text"
         />
 
-        <DividerGeneral class="divider" />
+        <DividerGeneral v-if="theme !== 'ftva'" class="divider" />
       </div>
     </div>
   </div>
@@ -85,110 +92,6 @@ const parsedContent = computed(() => {
   lang="scss"
   scoped
 >
-.section-media-with-text {
-  display: flex;
-  flex-direction: column;
-  flex-wrap: nowrap;
-  align-content: center;
-  align-items: center;
-  justify-content: flex-end;
-
-  background-color: var(--background-color);
-  width: 100%;
-
-  /* THEMES (used in section wrapper - zebra striping) */
-  --background-color: var(--color-theme, var(--color-white));
-
-  &.color-grey {
-    --background-color: var(--color-theme, var(--color-secondary-grey-02));
-  }
-
-  .flexible-page-block-container {
-    max-width: $container-l-main + px;
-    width: 100%;
-  }
-
-  >.meta {
-    width: 100%;
-  }
-
-  .visually-hidden {
-    @include visually-hidden;
-  }
-
-  .section-header {
-    margin-bottom: var(--space-xl);
-  }
-
-  .section-title {
-    @include step-3;
-    color: var(--color-primary-blue-03);
-    margin-bottom: var(--space-m);
-  }
-
-  .section-summary {
-    @include step-0;
-  }
-
-  .flexible-media-with-text {
-    margin-bottom: 56px;
-  }
-
-  :deep(.clipped-play) {
-    background-color: var(--background-color);
-  }
-
-  .divider {
-    max-width: $container-xl-full-width + px;
-    margin-bottom: 56px;
-  }
-
-  .meta:last-child {
-    .divider {
-      display: none;
-    }
-  }
-
-  // Breakpoints
-
-  @media #{$medium} {
-    .flexible-page-block-container {
-      padding: 0 var(--unit-gutter);
-    }
-
-    &.section-media-with-text {
-      padding-top: 40px;
-
-      .flexible-media-with-text {
-        margin-bottom: 32px;
-      }
-
-      .divider {
-        margin-bottom: 32px;
-      }
-    }
-  }
-
-  @media #{$small} {
-    .flexible-page-block-container {
-      padding: 0 var(--unit-gutter);
-    }
-
-    &.section-media-with-text {
-      padding-top: 32px;
-
-      .flexible-media-with-text {
-        margin-bottom: 24px;
-      }
-
-      .divider {
-        margin-bottom: 24px;
-      }
-
-      :deep(.clipped-play-mobile) {
-        background-color: var(--background-color);
-      }
-    }
-  }
-}
+@import "@/styles/default/_flexible-media-with-text.scss";
+@import "@/styles/ftva/_flexible-media-with-text.scss";
 </style>