+++ packages/vue-component-library/src/stories/Flexible_MediaWithText.stories.js
@@ -1,4 +1,6 @@
-import { mock, mockSingle } from './mock/Flexible_MediaWithText'
+import { computed } from 'vue'
+
+import { mock, mockFTVA, mockSingle } from './mock/Flexible_MediaWithText'
 import FlexibleMediaWithText from '@/lib-components/Flexible/MediaWithText'
 
 export default {
@@ -47,4 +49,23 @@ export function Single() {
 export const WithControls = Template.bind({})
 WithControls.args = {
   block: mock,
+}
+
+export function FTVADefault() {
+  return {
+    data() {
+      return { block: mockFTVA }
+    },
+    provide() {
+      return {
+        theme: computed(() => 'ftva'),
+      }
+    },
+    components: { FlexibleMediaWithText },
+    template: `
+        <flexible-media-with-text
+            :block="block"
+        />
+    `,
+  }
 }